[refactor] Set Controller
