+++ app/src/main/java/org/journey/android/di/ControllerModule.kt
@@ -8,6 +8,8 @@ import org.journey.android.login.controller.SignInController
 import org.journey.android.login.controller.SignInControllerImpl
 import org.journey.android.network.RetrofitInterface
 import org.journey.android.qualifier.UnAuthRetrofitService
+import org.journey.android.signup.controller.SignUpController
+import org.journey.android.signup.controller.SignUpControllerImpl
 import javax.inject.Singleton
 
 @Module
@@ -17,4 +19,9 @@ object ControllerModule {
     @Singleton
     fun provideSignInController(@UnAuthRetrofitService retrofitInterface: RetrofitInterface): SignInController =
         SignInControllerImpl(retrofitInterface)
+
+    @Provides
+    @Singleton
+    fun provideSignUpController(@UnAuthRetrofitService retrofitInterface: RetrofitInterface) : SignUpController =
+        SignUpControllerImpl(retrofitInterface)
 }
