+++ app/src/main/java/org/journey/android/signup/data/ResponseSocialSignUpDTO.kt
@@ -1,5 +1,4 @@
-package org.journey.android.signup.data.dto.response
-
+package org.journey.android.signup.data
 
 import com.google.gson.annotations.SerializedName
